# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chat Application API",
        "description": "API for managing chatrooms and messages.",
        "contact": {
            "name": "Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/chatrooms": {
            "get": {
                "tags": [
                    "Chatroom"
                ],
                "summary": "Get list of chatrooms",
                "description": "Retrieves all chatrooms.",
                "operationId": "getChatrooms",
                "responses": {
                    "200": {
                        "description": "List of chatrooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chatroom List"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Chatroom"
                ],
                "summary": "Create a new chatroom",
                "description": "Stores a newly created chatroom.",
                "operationId": "storeChatroom",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "General Chat"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A chatroom for general discussions"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chatroom created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chatroom Created"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chatrooms/{chatroom}/enter": {
            "post": {
                "tags": [
                    "Chatroom"
                ],
                "summary": "Enter a chatroom",
                "description": "Allows a user to enter a specified chatroom.",
                "operationId": "enterChatroom",
                "parameters": [
                    {
                        "name": "chatroom",
                        "in": "path",
                        "description": "ID of the chatroom to enter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully entered chatroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Entered Chatroom"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chatrooms/{chatroom}/leave": {
            "post": {
                "tags": [
                    "Chatroom"
                ],
                "summary": "Leave a chatroom",
                "description": "Allows a user to leave a specified chatroom.",
                "operationId": "leaveChatroom",
                "parameters": [
                    {
                        "name": "chatroom",
                        "in": "path",
                        "description": "ID of the chatroom to leave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully left chatroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Left Chatroom"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}